+++ packages/snap/src/permissions.ts
@@ -67,6 +67,7 @@ const metamaskPermissions = new Set([
   RpcRequestMethod.GetFeeForTransaction,
   // Client methods
   ClientRequestMethod.SignAndSendTransactionWithoutConfirmation,
+  ClientRequestMethod.SignProofOfOwnership,
 ]);
 
 const metamask = 'metamask';